+++ apps/mocksi-lite/content/EditMode/editMode.ts
@@ -2,7 +2,7 @@ import { MOCKSI_READONLY_STATE } from "../../consts";
 import type { ApplyAlteration } from "../Toast/EditToast";
 import { applyImageChanges } from "./actions";
 import { decorate } from "./decorator";
-import { Storage } from '../Storage';
+import { Storage } from '../utils/Storage';
 
 function openImageUploadModal(targetedElement: HTMLImageElement) {
 	// Create a container for the shadow DOM
